backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
bm25s==0.2.14
sentence-transformers==5.1.2
datasets
faiss-cpu==1.13.0
torch
transformers==4.57.1
nltk
accelerate==1.12.0